requirements.txt
numpy~=1.20.2
pandas~=1.1.5
Pillow~=8.2.0
matplotlib~=3.4.1